+++ src/main/java/net/silentchaos512/gear/util/GearHelper.java
@@ -256,7 +256,7 @@ public final class GearHelper {
     }
 
     private static boolean canBreakPermanently(ItemStack stack) {
-        return Config.Server.gearBreaksPermanently.get() || GearData.hasPart(stack, MiscUpgrades.RED_CARD.getPartId());
+        return Config.Common.gearBreaksPermanently.get() || GearData.hasPart(stack, MiscUpgrades.RED_CARD.getPartId());
     }
 
     public static boolean isBroken(ItemStack stack) {