+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- 'server' config is now 'common' and will appear in the config folder, like most configs do
+
 ## [1.10.2] - 2020-07-23
 ### Added
 - Racker trait (moving from Silent's Gems). It places netherrack.